+++ R/dependencies.R
@@ -445,10 +445,10 @@ is_readd_call <- function(expr){
   wide_deparse(expr[[1]]) %in% readd_fns
 }
 
-is_ignore <- function(expr){
+is_ignore_call <- function(expr){
   wide_deparse(expr[[1]]) %in% ignore_fns
 }
 
 ignore_pattern <- paste0(
-  "(", paste0(ignore_fns, collapse = "|"), ")\\s*\\([^\\)]*\\)"
+  "(\\s*|drake\\s*::\\s*|drake\\s*:::\\s*)ignore\\s*\\([^\\)]*\\)"
 )